move the client definition and default impl to a new file
